velocidadGenoma and minor fixes

+++ Client/Assets/Scripts/Bomb.cs
@@ -65,7 +65,7 @@ public class Bomb : MonoBehaviour
     */
     public IEnumerator CreateExplosions(Vector3 direction)
     {
-        int conversor = 100 / maxDistanceExplosion;
+        float conversor = 100 / maxDistanceExplosion;
         //Debug.Log(explosiondistanceGenoma / conversor);
         for (int i = 0; i < (explosiondistanceGenoma/conversor); i++)
         {
